% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_cna_seg.R
\name{annotate_cna_seg}
\alias{annotate_cna_seg}
\title{Annotate CN segments}
\usage{
annotate_cna_seg(seg.df, hmm.df)
}
\arguments{
\item{seg.df}{a data.frame with segment information}

\item{hmm.df}{a data.frame with bin information}
}
\value{
an annotated version of seg.df with a column wt.pv wit the pvalue of the
Wilcoxon test.
}
\description{
Annotate the CN segment predicted by the HMM. The signal in the segment is compared
to the signal in "neutral" segments nearby using a Wilcoxon test.
}
\author{
Jean Monlong
}
