% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_cna_call.R
\name{auto_cna_call}
\alias{auto_cna_call}
\title{Automated pipeline to call CNA}
\usage{
auto_cna_call(ge_df, comm_df, nb_metacells = 10, metacell_size = 3,
  trans_prob = 1e-04, baseline_cells = NULL,
  baseline_communities = NULL, prefix = "scCNAutils_out",
  nb_cores = 1, chrs = c(1:22, "X", "Y"), bin_mean_exp = 3,
  z_wins_th = 3, smooth_wsize = 3)
}
\arguments{
\item{ge_df}{normalized gene expression of all cells (e.g. output from
\code{\link{norm_ge}}.}

\item{comm_df}{a data.frame with community information, output from
\code{\link{find_communities}}.}

\item{nb_metacells}{the number of metacells per comunity.}

\item{metacell_size}{the number of cells in a metacell.}

\item{trans_prob}{the transition probability for the HMM.}

\item{baseline_cells}{cells to use as baseline.}

\item{baseline_communities}{communities to use as baseline. Used if
baseline.cells is NULL.}

\item{prefix}{the prefix to use for the files created by this function (e.g. graphs).}

\item{nb_cores}{the number of processors to use.}

\item{chrs}{the chromosome names to keep. NULL to include all the chromosomes.}

\item{bin_mean_exp}{the desired minimum mean expression in the bin.}

\item{z_wins_th}{the threshold to winsorize Z-score. Default is 3}

\item{smooth_wsize}{the window size for smoothing. Default is 3.}
}
\value{
a data.frame with CNAs
}
\description{
Automated pipeline to call CNA
}
\author{
Jean Monlong
}
