% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_cna_signal.R
\name{auto_cna_signal}
\alias{auto_cna_signal}
\title{Automated pipeline to compute CNA signal from scRNA expression}
\usage{
auto_cna_signal(data, genes_coord, prefix = "scCNAutils_out",
  nb_cores = 1, pause_after_qc = FALSE, use_cache = TRUE,
  sample_names = NULL, max_mito_prop = 0.2, min_total_exp = 0,
  cells_sel = NULL, chrs = c(1:22, "X", "Y"), cell_cycle = NULL,
  bin_mean_exp = 3, z_wins_th = 3, smooth_wsize = 3, cc_sd_th = 3,
  nb_pcs = 10, comm_k = 100)
}
\arguments{
\item{data}{a data.frame with gene expression or the path to the folder with the
'matrix.mtx', 'genes.tsv' and 'barcodes.tsv' files. A list if multiple samples.}

\item{genes_coord}{either a file name or a data.frame with coordinates
and gene names.}

\item{prefix}{the prefix to use for the files created by this function (e.g. graphs).}

\item{nb_cores}{the number of processors to use.}

\item{pause_after_qc}{pause after the QC to pick custom QC thresholds.}

\item{use_cache}{should intermediate files used and avoid redoing steps?}

\item{sample_names}{the names of each sample. If NULL, tries to use data's names.}

\item{max_mito_prop}{the maximum proportion of mitochondrial RNA.}

\item{min_total_exp}{the minimum total cell expression}

\item{cells_sel}{consider only these cells. Other cells filtered no matter what.}

\item{chrs}{the chromosome names to keep. NULL to include all the chromosomes.}

\item{cell_cycle}{if non-null, either a file or data.frame to compute cell
cycle scores. See details.}

\item{bin_mean_exp}{the desired minimum mean expression in the bin.}

\item{z_wins_th}{the threshold to winsorize Z-score. Default is 3}

\item{smooth_wsize}{the window size for smoothing. Default is 3.}

\item{cc_sd_th}{the number of SD used for the thresholds when defining cycling cells.}

\item{nb_pcs}{the number of PCs used in the community detection or tSNE.}

\item{comm_k}{the number of nearest neighbor for the KNN graph. Default 100.}
}
\value{
a data.frame with QC, community and tSNE for each cell.
}
\description{
Goes from reading raw gene counts to CNA-level signal, tSNE and community detection.
}
\author{
Jean Monlong
}
