% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_genes.R
\name{bin_genes}
\alias{bin_genes}
\title{Merge consecutive genes into expressed bins}
\usage{
bin_genes(ge_df, mean_exp = 3, nb_cores = 1, rcpp = FALSE)
}
\arguments{
\item{ge_df}{the input gene expression with coordinate columns (chr, start, end)
and then one column per cell.}

\item{mean_exp}{the desired minimum mean expression in the bin.}

\item{nb_cores}{the number of processors to use.}

\item{rcpp}{use Rcpp function. Default is FALSE. More memory-efficient and
faster when running on one core.}
}
\value{
a data.frame with bin expression.
}
\description{
Merge consecutive genes into expressed bins
}
\author{
Jean Monlong
}
