% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_cna.R
\name{call_cna}
\alias{call_cna}
\title{Call CNA}
\usage{
call_cna(z_df, trans_prob = 1e-04, nb_cores = 1, mc_info = NULL)
}
\arguments{
\item{z_df}{the Z-scores, from \code{\link{zscore}}.}

\item{trans_prob}{the transition probability for the HMM.}

\item{nb_cores}{the number of processor to use.}

\item{mc_info}{the information about the metacells, if relevant. Default is NULL.}
}
\value{
a data.frame with the CNA calls.
}
\description{
Calls CNA using a HMM approach.
}
\author{
Jean Monlong
}
