% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_cna_multisamps.R
\name{call_cna_multisamps}
\alias{call_cna_multisamps}
\title{Call CNA}
\usage{
call_cna_multisamps(z_df, mc_info, trans_prob = 1e-04, nb_cores = 1)
}
\arguments{
\item{z_df}{the Z-scores, from \code{\link{zscore}}.}

\item{mc_info}{the information about the metacells, if relevant. Default is NULL.}

\item{trans_prob}{the transition probability for the HMM.}

\item{nb_cores}{the number of processor to use.}
}
\value{
a data.frame with the CNA calls.
}
\description{
Calls CNA using a HMM approach considering multiple samples at
the same time.
}
\author{
Jean Monlong
}
