% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaHMM.R
\name{cnaHMM}
\alias{cnaHMM}
\title{Call CNA using a HMM approach}
\usage{
cnaHMM(
  ge,
  trans.prob = 1e-04,
  log.mu = TRUE,
  max.samps = 200,
  perChr = TRUE,
  winsorize = TRUE
)
}
\arguments{
\item{ge}{the input gene expression}

\item{trans.prob}{the transition probability}

\item{log.mu}{should the mu be log-scaled?}

\item{max.samps}{the maximum number of samples accepted.}

\item{perChr}{run on each chromosome separately?}

\item{winsorize}{winsorize zscores to limit the effect of extreme outliers.}
}
\value{
a data.frame with info about CNA segments
}
\description{
Call CNA using a HMM approach
}
\author{
Jean Monlong
}
\keyword{internal}
