% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_coord.R
\name{convert_to_coord}
\alias{convert_to_coord}
\title{Convert gene symbols to coordinates}
\usage{
convert_to_coord(ge_df, genes_coord, chrs = c(1:22, "X", "Y"),
  rm_dup = TRUE)
}
\arguments{
\item{ge_df}{the data.frame with gene expression and one column 'symbol'
with gene names.}

\item{genes_coord}{either a file name or a data.frame with coordinates
and gene names.}

\item{chrs}{the chromosome names to keep. NULL to include all the
chromosomes.}

\item{rm_dup}{remove duplicated coordinates? Default is TRUE.}
}
\value{
a data.frame with columns 'chr', 'start', 'end' columns with genes
coordinates (and still one column per barcode).
}
\description{
Convert the 'symbol' column (gene names) into three columns with gene coordinates
'chr', 'start' and 'end'.
}
\details{
If \emph{genes_coord} is a filename, the file is expected to be a tab-delimited
file with four columns: 'chr', 'start', 'end', 'symbol'. The order of the
columns is not important.

The gene names in column 'symbol' should match the gene names in the input
\emph{ge_df}.
}
\author{
Jean Monlong
}
