% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_cycling_cells.R
\name{define_cycling_cells}
\alias{define_cycling_cells}
\title{Define cycling cells}
\usage{
define_cycling_cells(qc_df, sd_th = 3)
}
\arguments{
\item{qc_df}{the output data.frame from \code{\link{qc_cells}} (ran with a
non-null \emph{cell_cycle} parameter)}

\item{sd_th}{the number of SD used for the thresholds.}
}
\value{
a list with
\item{cells.noc}{a vector with the names of non-cycling cells}
\item{graphs}{a list of ggplot2 graphs}
}
\description{
Define cycling cells
}
\author{
Jean Monlong
}
