% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_communities.R
\name{find_communities}
\alias{find_communities}
\title{Community detection}
\usage{
find_communities(pca_o, nb_pcs = 10, k = 100)
}
\arguments{
\item{pca_o}{the output of \code{\link{run_pca}}}

\item{nb_pcs}{the number of PCs to use. Default 10.}

\item{k}{the number of nearest neighbor for the KNN graph. Default 100.}
}
\value{
a data.frame with two columns: 'cell' and 'community'.
}
\description{
Build a KNN graph and run Louvain algorithm for community detection.
}
\author{
Jean Monlong
}
