% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_info.R
\name{gene_info}
\alias{gene_info}
\title{Gene information}
\usage{
gene_info(ge_df, genes_coord, subset_cells = 10000)
}
\arguments{
\item{ge_df}{a data.frame with gene expression
information across cells. E.g. after \code{norm.ge}.}

\item{genes_coord}{either a file name or a data.frame with coordinates
and gene names.}

\item{subset_cells}{the maximum number of cells to use.}
}
\value{
a data.frame with summary stats for each gene.
}
\description{
Computes information at the gene level that can be used to annotate
the results later (e.g. CNA calls).
}
\details{
If the input data.frame has a symbol column, 'genes_coord' is used to
add the coordinates, later used to overlap with region/CNA calls. If
the input data.frame has coordinates, 'genes_coord' is used to retrieve
the gene name. In both cases a row in the input data.frame must be a
gene, not a bin.

It's better to run this function after normalization (\code{norm_ge}
but before binning (\code{bin_genes}).
}
\author{
Jean Monlong
}
