% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metacells.R
\name{make_metacells}
\alias{make_metacells}
\title{Make metacells}
\usage{
make_metacells(
  ge_df,
  comm_df,
  nb_metacells = 10,
  metacell_size = 3,
  baseline_cells = NULL,
  nb_cores = 1,
  max_baseline_comm = 3
)
}
\arguments{
\item{ge_df}{gene expression of all cells}

\item{comm_df}{a data.frame with community information, output from
\code{\link{find_communities}}.}

\item{nb_metacells}{the number of metacells per comunity.}

\item{metacell_size}{the number of cells in a metacell.}

\item{baseline_cells}{the cells to use for baseline communities.}

\item{nb_cores}{the number of processor to use.}

\item{max_baseline_comm}{the maximum number of baseline communities to generate.}
}
\value{
a list with
\item{ge}{a data.frame with coordinates and gene expression for each metacell.}
\item{info}{information about which metacell correspond to which community.}
\item{mc_cells}{information about which cells were used for each metacell.}
}
\description{
Randomly select cells in each community and merge them to create metacells
with higher resolution.
}
\author{
Jean Monlong
}
