% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_samples.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge expression of multiple samples}
\usage{
merge_samples(ge_list, sample_names = NULL)
}
\arguments{
\item{ge_list}{a list of ge_df (e.g. read from \code{\link{read_mtx}}).}

\item{sample_names}{the names of each sample. If NULL, tries to use ge_list's names.}
}
\value{
a list with
\item{ge}{the merged gene expression data.frame}
\item{info}{a data.frame with new and original cell names, and
corresponding sample name}
}
\description{
The expression of multiple samples are merged. New cell names are produced as
SAMPLE_CELL.
}
\author{
Jean Monlong
}
