% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells_cluster.R
\name{metacells_cluster}
\alias{metacells_cluster}
\title{Metacells by clustering within communities.}
\usage{
metacells_cluster(
  groups.df,
  ge_df,
  metacell_size,
  nb_cores = 1,
  hclust_method = "ward.D"
)
}
\arguments{
\item{groups.df}{the filtered data.frame with community information}

\item{ge_df}{normalized gene expression of all cells (e.g. output from
\code{\link{norm_ge}}.}

\item{metacell_size}{the number of cells in a metacell.}

\item{nb_cores}{the number of processor to use.}

\item{hclust_method}{the 'method' used when clustering (see hclust).}
}
\value{
a list with
\item{cells.tm}{a list with cell names for each metacell.}
\item{cells.info}{a data.frame recording the community of origina for each metacell.}
}
\description{
Metacells by clustering within communities.
}
\author{
Jean Monlong
}
\keyword{internal}
