% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells_subsample.R
\name{metacells_subsample}
\alias{metacells_subsample}
\title{Metacells by subsampling within communities.}
\usage{
metacells_subsample(groups.df, nb_metacells, metacell_size)
}
\arguments{
\item{groups.df}{the filtered data.frame with community information}

\item{nb_metacells}{the number of metacells per comunity.}

\item{metacell_size}{the number of cells in a metacell.}
}
\value{
a list with
\item{cells.tm}{a list with cell names for each metacell.}
\item{cells.info}{a data.frame recording the community of origina for each metacell.}
}
\description{
Metacells by subsampling within communities.
}
\author{
Jean Monlong
}
\keyword{internal}
