% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_ge_tmm.R
\name{norm_ge_tmm}
\alias{norm_ge_tmm}
\title{Normalize gene expression using TMM}
\usage{
norm_ge_tmm(ge_df, nb_cores = 1)
}
\arguments{
\item{ge_df}{the input gene expression.}

\item{nb_cores}{the number of processors to use.}
}
\value{
a data.frame with the normalized expression.
}
\description{
Normalize gene expression using TMM
}
\author{
Jean Monlong
}
\keyword{internal}
