% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cna.R
\name{plot_cna}
\alias{plot_cna}
\title{Heatmap of CNA}
\usage{
plot_cna(cna_df, chrs_order = c(1:22, "X", "Y"))
}
\arguments{
\item{cna_df}{CNA from \code{\link{call_cna}}.}

\item{chrs_order}{order of the chroosomes in the graph.}
}
\value{
a ggplot2 graph
}
\description{
Heatmap of CNA
}
\author{
Jean Monlong
}
