% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_communities.R
\name{plot_communities}
\alias{plot_communities}
\title{Community graphs}
\usage{
plot_communities(comm_df, qc_df = NULL)
}
\arguments{
\item{comm_df}{the output data.frame from \code{\link{find_communities}}}

\item{qc_df}{a data.frame with QC metrics (output from \code{\link{qc_cells}}).
Default is NULL (i.e. not used)}
}
\value{
a list of ggplot2 graphs.
}
\description{
Community graphs
}
\author{
Jean Monlong
}
