% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qc_cells.R
\name{plot_qc_cells}
\alias{plot_qc_cells}
\title{QC graphs}
\usage{
plot_qc_cells(qc_df, info_df = NULL)
}
\arguments{
\item{qc_df}{the output data.frame from \code{\link{qc_cells}}}

\item{info_df}{a data.frame with sample merge info (output from
\code{\link{merge_samples}}). Default is NULL (i.e. not used)}
}
\value{
a list of ggplots
}
\description{
QC graphs
}
\examples{
\dontrun{
ggp.l = plot_qc_cells(qc_df)

## Print first graph
ggpl.l[[1]]

## Customize ggplot
ggpl.l[[1]] + ggtitle('First QC graph')
}
}
\author{
Jean Monlong
}
