% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsne.R
\name{plot_tsne}
\alias{plot_tsne}
\title{tSNE graphs}
\usage{
plot_tsne(tsne_df, qc_df = NULL)
}
\arguments{
\item{tsne_df}{the output data.frame from \code{\link{run_tsne}} (columns: cell, tsne1, tsne2)}

\item{qc_df}{a data.frame with QC metrics (output from \code{\link{qc_cells}}). Default is NULL
(i.e. not used)}
}
\value{
a list of ggplot objects
}
\description{
tSNE graphs
}
\author{
Jean Monlong
}
