% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{UMAP graphs}
\usage{
plot_umap(umap_df, qc_df = NULL, comm_df = NULL, info_df = NULL)
}
\arguments{
\item{umap_df}{the output data.frame from \code{\link{run_umap}} (columns: cell, umap1, umap2)}

\item{qc_df}{a data.frame with QC metrics (output from \code{\link{qc_cells}}).
Default is NULL (i.e. not used)}

\item{comm_df}{a data.frame with communities (output from \code{\link{find_communities}}).
Default is NULL (i.e. not used)}

\item{info_df}{a data.frame with sample merge info (output from
\code{\link{merge_samples}}).}
}
\value{
a list of ggplot objects
}
\description{
UMAP graphs colored according to QC metrics or sample labels.
}
\details{
If the QC data.frame is provided, the distribution of QC metrics is
shown to investigate if some communities are batch effects.

If multiple samples were merged (\code{\link{merge_samples}}), the
points can be colored by sample of origin by providing the info_df
data.frame.

If any qc_df/comm_df/info_df are null but their columns present in umap_df,
their corresponding graphs will be generated. Hence a merged version of
umap_df, comm_df, qc_df and info_df works (e.g. output of
\code{\link{auto_cna_signal}}.
}
\examples{
\dontrun{
ggp.l = plot_umap(umap_df, qc_df, comm_df)

## Print first graph
ggpl.l[[1]]

## Customize ggplot
ggpl.l[[1]] + ggtitle('First umap graph')
}
}
\author{
Jean Monlong
}
