% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_cells.R
\name{qc_cells}
\alias{qc_cells}
\title{Compute quality control metrics for each cell}
\usage{
qc_cells(ge_df, cell_cycle = NULL)
}
\arguments{
\item{ge_df}{the input gene expression with a 'symbol' column and then one
column per cell.}

\item{cell_cycle}{if non-null, either a file or data.frame to compute cell
cycle scores. See details.}
}
\value{
a data.frame with qc metrics per cell.
}
\description{
If cell_cycle is provided it should be a data.frame (or a tsv file) with
two columns: 'symbol' with gene names, and 'phase' with the cell cycle phase
(e.g. either 'G1.S' or 'G2.M').
}
\author{
Jean Monlong
}
