% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mtx.R
\name{read_mtx}
\alias{read_mtx}
\title{Read a trio of genes, barcodes and mtx files.}
\usage{
read_mtx(
  mtx_file = "matrix.mtx",
  genes_file = "genes.tsv",
  barcodes_file = "barcodes.tsv",
  path = ".",
  rm_dup = TRUE,
  genes_col = 2,
  min_barcode_exp = 1,
  filter_zero_genes = TRUE
)
}
\arguments{
\item{mtx_file}{the path to the mtx file}

\item{genes_file}{the path to the genes file.}

\item{barcodes_file}{the path to the barcodes file}

\item{path}{the path to the folder containing the files}

\item{rm_dup}{remove duplicated gene names? Default is TRUE.}

\item{genes_col}{the column to use in genes_file. Default is 2.}

\item{min_barcode_exp}{minimum total expression in barcodes/cells. Default is 1.}

\item{filter_zero_genes}{filter genes with no reads in any cells? Default is TRUE.}
}
\value{
a data.frame with a 'symbol' column with gene names and one
column per barcode.
}
\description{
Read a trio of genes, barcodes and mtx files.
}
\author{
Jean Monlong
}
