% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebin_cov.R
\name{rebin_cov}
\alias{rebin_cov}
\title{Re-bin coverage data}
\usage{
rebin_cov(cov_df, bins)
}
\arguments{
\item{cov_df}{a data.frame with coverage information.}

\item{bins}{a data.frame or GRanges object with the new bins.}
}
\value{
a data.frame with new bins
}
\description{
The new bins are overlapped with the regions in cov_df to compute a weight.
The coverage in the new bin is the weighted sum of the coverage in overlapping
regions of cov_df.
}
\details{
Coordinates in data.frame are expected to be defined by columns names 'chr',
'start' and 'end'.
}
\author{
Jean Monlong
}
