% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_cv_outliers.R
\name{rm_cv_outliers}
\alias{rm_cv_outliers}
\title{Remove outliers based on the coefficient of variation}
\usage{
rm_cv_outliers(ge_df, ol_quant_th = 0.99, ol_sd_th = 5)
}
\arguments{
\item{ge_df}{the expression data.frame.}

\item{ol_quant_th}{the quantile threshold. Default is 0.99 (removes the
top 1\% with highest values).}

\item{ol_sd_th}{the SD-based treshold. Default is 5.}
}
\value{
a subset of the ge_df data.frame
}
\description{
Compute the coefficient of variation for each gene/bin and remove the ones
with the highest values, either based on a quantile or SD-based threshold.
The genes/bins that satisfy both quantile and SD-based thresholds are
removed.
}
\author{
Jean Monlong
}
