% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_louvain.R
\name{run_louvain}
\alias{run_louvain}
\title{Python wrapper to run Louvain}
\usage{
run_louvain(graph, gamma = 1, nreps = 1, nb_cores = 1)
}
\arguments{
\item{graph}{a igraph object}

\item{gamma}{a vector of gamma. Default 1.}

\item{nreps}{the number of repetition for each gamma, Default 1.}

\item{nb_cores}{the number of processors to use. Default is 1.}
}
\value{
a list with
\item{comm}{a  data.frame with the community for each gamma}
\item{gamma}{the input gammas corresponding to the columns of comm}
}
\description{
Louvain on an igraph object.
}
\details{
This functions depends on Python and louvain being installed. Make sure igraph,
louvain and numpy are installed. For example with something like:
'pip install python-igraph louvain numpy'.
}
\author{
Jean Monlong
}
