% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{run_pca}
\alias{run_pca}
\title{Run PCA}
\usage{
run_pca(z_df, core_cells = NULL, out_pcs = 100)
}
\arguments{
\item{z_df}{a data.frame with z-scores for each cell}

\item{core_cells}{if non-NULL, a vector with the names of the cells to use as core
cells. See details. Default is NULL.}

\item{out_pcs}{the number of top PCs to report. Default is 100.}
}
\value{
a list with
\item{x}{the PC matrix}
\item{sdev}{the standard deviations of the PCs}
\item{sdev.graph}{a ggplot graph of the sdev}
}
\description{
PCA analysis, eventually using a subset of core cells for the PC construction.
}
\details{
Cells in core_cells are used to build the principal components to which
all cells are then projected to. Usually used to reduce the effect of
cell cycle in the PCA, by using only cells that don't cycle (see \code{\link{qc_cells}})
as \emph{core_cells}.

The graph (\emph{sdev.graph}) shows the standard deviation for the top 50 PCs. To
show more/less PCs, add \code{xlim(1,N)} to the \emph{sdev.graph}. See examples.
}
\examples{
\dontrun{
pca.o = run_pca(z)

## Zoom in to the top 20 PCs
pca.o$sdev.graph + xlim(1,20)
}
}
\author{
Jean Monlong
}
