% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tsne.R
\name{run_tsne}
\alias{run_tsne}
\title{Run tSNE}
\usage{
run_tsne(pca_o, nb_pcs = 10, nb_it = 1000)
}
\arguments{
\item{pca_o}{the output of \code{\link{run_pca}}}

\item{nb_pcs}{the number of PCs to use. Default 10.}

\item{nb_it}{the number of iterations. Default 1000.}
}
\value{
a data.frame with columns: cell, tsne1, tsne2
}
\description{
Run tSNE
}
\author{
Jean Monlong
}
