% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_umap.R
\name{run_umap}
\alias{run_umap}
\title{Run UMAP}
\usage{
run_umap(pca_o, nb_pcs = 10, nb_neighbors = 5)
}
\arguments{
\item{pca_o}{the output of \code{\link{run_pca}}}

\item{nb_pcs}{the number of PCs to use. Default 10.}

\item{nb_neighbors}{the number of neighbors. Default 5.}
}
\value{
a data.frame with columns: cell, umap1, umap2
}
\description{
UMAP on the PCA results.
}
\details{
This functions depends on Python and UMAP being installed. Make sure umap-mearn,
sklearn, numpy, scipy and pandas are installed. For example with something like:
'pip install sklearn numpy scipy pandas umap-learn'.
}
\author{
Jean Monlong
}
