% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCNAutils.R
\docType{package}
\name{scCNAutils-package}
\alias{scCNAutils}
\alias{scCNAutils-package}
\title{scCNAutils}
\description{
A cleaned up version of the functions I used to analyze copy number aberrations in single-cell data. Still in development.
}
\author{
\strong{Maintainer}: Jean Monlong \email{jean.monlong@gmail.com}

}
