% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_movingw.R
\name{smooth_movingw}
\alias{smooth_movingw}
\title{Moving-window smoothin}
\usage{
smooth_movingw(df, wsize = 3, nb_cores = 1, FUN = stats::median)
}
\arguments{
\item{df}{the input data.frame with coordinate columns (chr, start, end) and then
one column per cell}

\item{wsize}{the window size. Default is 3.}

\item{nb_cores}{the number of processors to use.}

\item{FUN}{the function to apply to each window. Default is median.}
}
\value{
a data.frame with smoothed signal.
}
\description{
Moving-window smoothin
}
\author{
Jean Monlong
}
