% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smoothMovingC}
\alias{smoothMovingC}
\title{Smooth signal}
\usage{
smoothMovingC(ge, winsize)
}
\arguments{
\item{ge}{an ordered gene expression matrix}

\item{winsize}{the size of the sliding window (assumed to be odd).}
}
\description{
Smooth signal
}
