% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tmmNormC}
\alias{tmmNormC}
\title{Compute the normalization factor}
\usage{
tmmNormC(ge, cont)
}
\arguments{
\item{ge}{gene expression amtrix}

\item{cont}{index of the control sample. R index so starting at 1.}
}
\description{
Compute the normalization factor
}
