% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_browser.R
\name{tsne_browser}
\alias{tsne_browser}
\title{Shiny application to visualize tSNE results}
\usage{
tsne_browser(cells_df, nb_points = 5000, plot_dim = 800)
}
\arguments{
\item{cells_df}{the data.frame with tSNE and other information for each cell}

\item{nb_points}{the default number of points to show. See details.}

\item{plot_dim}{the dimension of the plot in pixels.}
}
\value{
opens a Shiny app in a web-browser.
}
\description{
Interactive application to visualize the tSNE results: zoom, hover
information, different colors.
}
\details{
Drawing thousands of points in a web-browser can be demanding. To reduce the
number of points (cells) to draw, close-by cells are merged into bigger
points. The merging is done separately for different samples/communities to
be able to color them if necessary. The user can decide how many points to
draw with the 'nb_points' parameter or directly within the application.
In practice, increase the number of points until the app gets too slow.
}
\author{
Jean Monlong
}
