% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsor.R
\name{winsor}
\alias{winsor}
\title{Winsorize}
\usage{
winsor(x, u = NULL, l = NULL, qu = NULL)
}
\arguments{
\item{x}{input vector}

\item{u}{upper limit}

\item{l}{lower limit}

\item{qu}{upper quartile, if 'u' is null}
}
\value{
winsorized vector
}
\description{
Winsorize
}
\author{
Jean Monlong
}
\keyword{internal}
