% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_cna.R
\name{annotate_cna}
\alias{annotate_cna}
\title{Annotate CNAs with gene information}
\usage{
annotate_cna(cna_df, gene_info)
}
\arguments{
\item{cna_df}{a data.frame with CNA calls. E.g. data.frames created
by \code{call_cna_*} functions.}

\item{gene_info}{a data.frame with gene information created by the
\code{gene_info} function}
}
\value{
the input data.frame with two new columns with all/expressed genes.
}
\description{
Add columns with the names of the genes in the region, total
and expressed genes only.
}
\details{
The subset of "expressed" genes is made out of genes with
non-zero expression in at least 10% of cells or with a mean
expression higher than 0.5.
}
\author{
Jean Monlong
}
