% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_ge.R
\name{norm_ge}
\alias{norm_ge}
\title{Normalize gene expression}
\usage{
norm_ge(ge_df, method = c("tmm", "total"), nb_cores = 1,
  rcpp = FALSE)
}
\arguments{
\item{ge_df}{the input gene expression}

\item{method}{the normalization method}

\item{nb_cores}{the number of processors to use.}

\item{rcpp}{use Rcpp function. Default is FALSE. More memory-efficient and
faster when running on one core.}
}
\value{
a data.frame with the normalized expression.
}
\description{
The expression of each cell is normalized to account for depth differences.
}
\author{
Jean Monlong
}
