% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_ge_total.R
\name{norm_ge_total}
\alias{norm_ge_total}
\title{Normalize gene expression using total expression}
\usage{
norm_ge_total(ge_df, nb.cores = 1)
}
\arguments{
\item{ge_df}{the input gene expression}

\item{nb.cores}{the number of processors to use.}
}
\value{
a data.frame with the normalized expression.
}
\description{
Normalize gene expression using total expression
}
\author{
Jean Monlong
}
\keyword{internal}
