% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qc_cells.R
\name{plot_qc_cells}
\alias{plot_qc_cells}
\title{QC graphs}
\usage{
plot_qc_cells(qc_df)
}
\arguments{
\item{qc_df}{the output data.frame from \code{\link{qc_cells}}}
}
\value{
a list of ggplots
}
\description{
QC graphs
}
\author{
Jean Monlong
}
