% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_filter.R
\name{qc_filter}
\alias{qc_filter}
\title{Filter cells based from QC results}
\usage{
qc_filter(ge_df, qc_df, max_mito_prop = 0.2, min_total_exp = 0)
}
\arguments{
\item{ge_df}{the input gene expression with a 'symbol' column and then one
column per cell.}

\item{qc_df}{the output data.frame from qc_cells}

\item{max_mito_prop}{the maximum proportion of mitochondrial RNA.}

\item{min_total_exp}{the minimum total cell expression}
}
\value{
ge_df with only the cells that passed the filters
}
\description{
Filter cells based from QC results
}
\author{
Jean Monlong
}
