% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_movingw.R
\name{smooth_movingw}
\alias{smooth_movingw}
\title{Moving-window smoothing}
\usage{
smooth_movingw(df, wsize = 3, nb_cores = 1, FUN = stats::median,
  rcpp = FALSE)
}
\arguments{
\item{df}{the input data.frame with coordinate columns (chr, start, end) and then
one column per cell}

\item{wsize}{the window size. Default is 3.}

\item{nb_cores}{the number of processors to use.}

\item{FUN}{the function to apply to each window. Default is median.}

\item{rcpp}{use Rcpp function. Default is FALSE. More memory-efficient and
faster when running on one core.}
}
\value{
a data.frame with smoothed signal.
}
\description{
The expression/score of a gene/bin is replaced by a summary of bins around.
For example the median across 3 bins.
}
\author{
Jean Monlong
}
