% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsor.R
\name{winsor}
\alias{winsor}
\title{Winsorize}
\usage{
winsor(x, u = NULL, l = NULL, uq = NULL)
}
\arguments{
\item{x}{input vector}

\item{u}{upper limit}

\item{l}{lower limit}

\item{uq}{the quantile for the upper limit. Used is u is NULL.}
}
\value{
winsorized vector
}
\description{
Convenience function to winsorize a vector.
}
\author{
Jean Monlong
}
