% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Compute Z-score}
\usage{
zscore(ge_df, wins_th = 3, method = c("z", "norm"), normals = NULL)
}
\arguments{
\item{ge_df}{the input expression data.frame}

\item{wins_th}{the threshold to winsorize Z-score. Default is 3}

\item{method}{the normalization method. Either 'z' or norm'.}

\item{normals}{the cells to use as normals. If NULL (default) all cells are used
as normals}
}
\value{
a data.frame with Z-scores.
}
\description{
Compute Z-score
}
\author{
Jean Monlong
}
